doing int instead of float for some gradients ... doesnt really help.

+++ piggyHOG_src/PgHog.cpp
@@ -69,13 +69,13 @@ inline void PgHog::gradient(int x, int y, Mat img, Mat &oriImg, Mat &magImg){
     x = clamp(x, 1, img.cols-1);
     y = clamp(y, 1, img.rows-1);
 
-    float gradX;
-    float gradY;
+    int gradX;
+    int gradY;
     float max_mag = 0.0f; //TODO: check range ... can this be uchar?
 
     for(int channel=0; channel<3; channel++){
-        float tmp_gradX = img.at<cv::Vec3b>(y,x+1)[channel] - img.at<cv::Vec3b>(y,x-1)[channel];
-        float tmp_gradY = img.at<cv::Vec3b>(y+1,x)[channel] - img.at<cv::Vec3b>(y-1,x)[channel];
+        int tmp_gradX = img.at<cv::Vec3b>(y,x+1)[channel] - img.at<cv::Vec3b>(y,x-1)[channel];
+        int tmp_gradY = img.at<cv::Vec3b>(y+1,x)[channel] - img.at<cv::Vec3b>(y-1,x)[channel];
 
         //indexing the data directly instead of using .at -- doesn't help with perf.
         //float tmp_gradX = img.data[y*img.rows*3 + (x+1)*3 + channel] - img.data[y*img.rows*3 + (x-1)*3 + channel];
